requirements.txt
# Core runtime requirements for AI-OS prototype
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0
PyYAML>=6.0
rich>=13.0
prometheus-client>=0.17.0
keyring>=23.0.0
# AI-OS Python Requirements
# Core agent dependencies

# LLM Integration
openai>=1.0.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-community>=0.0.10

# Speech Recognition & Synthesis
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
whisper>=1.0.0

# Computer Vision (for gesture recognition)
opencv-python>=4.8.0
mediapipe>=0.10.0

# System Integration
psutil>=5.9.0
watchdog>=3.0.0
pynput>=1.7.6

# UI & Terminal
rich>=13.0.0
prompt-toolkit>=3.0.0
textual>=0.40.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.0.0

# Networking
httpx>=0.25.0
websockets>=12.0

# Logging & Monitoring
structlog>=23.0.0
